# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "GoalSphere API",
    "description": "API for GoalSphere - Arabic Football News Platform",
    "version": "1.0.0",
    "contact": {
      "name": "GoalSphere Team",
      "email": "api@goalsphere.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Development server"
    },
    {
      "url": "https://goalsphere.com/api",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/ingestion/start": {
      "post": {
        "tags": ["Content Ingestion"],
        "summary": "Start content ingestion process",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "source": {
                    "type": "string",
                    "enum": ["sportmonks", "api_football", "web_scraping"],
                    "description": "Data source"
                  },
                  "categories": {
                    "type": "array",
                    "items": { "type": "string" },
                    "description": "Categories to fetch"
                  },
                  "force_refresh": {
                    "type": "boolean",
                    "description": "Force refresh existing data"
                  }
                },
                "required": ["source"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Ingestion started successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "job_id": { "type": "string" },
                    "results": {
                      "type": "object",
                      "properties": {
                        "processed": { "type": "integer" },
                        "created": { "type": "integer" },
                        "updated": { "type": "integer" },
                        "errors": { "type": "array", "items": { "type": "string" } }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Ingestion failed"
          }
        }
      }
    },
    "/ingestion/status": {
      "get": {
        "tags": ["Content Ingestion"],
        "summary": "Get ingestion job status",
        "parameters": [
          {
            "name": "job_id",
            "in": "query",
            "description": "Specific job ID",
            "schema": { "type": "string" }
          },
          {
            "name": "source",
            "in": "query",
            "description": "Filter by source",
            "schema": { "type": "string" }
          },
          {
            "name": "status",
            "in": "query",
            "description": "Filter by status",
            "schema": { "type": "string", "enum": ["pending", "running", "completed", "failed"] }
          }
        ],
        "responses": {
          "200": {
            "description": "Job status retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/IngestionJob" }
                    },
                    "count": { "type": "integer" }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "IngestionJob": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "source": { "type": "string" },
          "status": { "type": "string", "enum": ["pending", "running", "completed", "failed"] },
          "started_at": { "type": "string", "format": "date-time" },
          "completed_at": { "type": "string", "format": "date-time" },
          "articles_processed": { "type": "integer" },
          "articles_created": { "type": "integer" },
          "articles_updated": { "type": "integer" },
          "errors": { "type": "array", "items": { "type": "string" } },
          "created_at": { "type": "string", "format": "date-time" }
        }
      }
    }
  }
}